app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
redis
httpx
pydantic
firebase-admin
